continue

+++ parser/src/token_set.rs
@@ -13,11 +13,11 @@ use kitty_syntax::TokenKind;
 // This implementation is mostly stolen from rust-analyzer:
 // https://github.com/rust-analyzer/rust-analyzer/blob/b73b321478d3b2a98d380eb79de717e01620c4e9/crates/parser/src/token_set.rs
 #[derive(Debug, Clone, Copy, Default)]
-pub(crate) struct TokenSet(u64);
+pub(crate) struct TokenSet(u128);
 
 impl TokenSet {
-    pub(crate) const ALL: Self = Self(u64::MAX);
-    pub(crate) const NONE: Self = Self(u64::MIN);
+    pub(crate) const ALL: Self = Self(u128::MAX);
+    pub(crate) const NONE: Self = Self(u128::MIN);
 
     pub(crate) const fn new<const LEN: usize>(kinds: [TokenKind; LEN]) -> Self {
         let mut value = 0;
@@ -44,7 +44,7 @@ impl TokenSet {
     }
 }
 
-const fn mask(kind: TokenKind) -> u64 {
+const fn mask(kind: TokenKind) -> u128 {
     1 << (kind as usize)
 }
